backend/pyproject.toml
[project]
name = "oxytec-feasibility-platform"
version = "0.1.0"
description = "Multi-Agent Feasibility Study Platform for Oxytec"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy[asyncio]>=2.0.23",
    "asyncpg>=0.29.0",
    "alembic>=1.12.1",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "anthropic>=0.7.0",
    "langgraph>=0.0.20",
    "langchain>=0.1.0",
    "langchain-anthropic>=0.1.0",
    "aiohttp>=3.9.1",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "pgvector>=0.2.4",
    "openai>=1.6.0",
    "pymupdf>=1.23.8",
    "python-docx>=1.1.0",
    "pandas>=2.1.4",
    "openpyxl>=3.1.2",
    "structlog>=23.2.0",
    "httpx>=0.25.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "black>=23.12.0",
    "ruff>=0.1.8",
    "mypy>=1.7.1",
    "locust>=2.19.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
